+++ src/main/java/phanes/replay/comment/service/CommentService.java
@@ -3,9 +3,9 @@ package phanes.replay.comment.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import phanes.replay.comment.dto.mapper.CommentMapper;
 import phanes.replay.comment.dto.request.CommentCreateRq;
 import phanes.replay.comment.dto.response.CommentRs;
-import phanes.replay.comment.mapper.CommentMapper;
 import phanes.replay.exception.CommentNotFoundException;
 import phanes.replay.gathering.domain.Gathering;
 import phanes.replay.gathering.domain.GatheringComment;
