+++ src/main/java/phanes/replay/review/persistence/repository/ReviewImageRepository.java
@@ -1,4 +1,4 @@
-package phanes.replay.review.repository;
+package phanes.replay.review.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import phanes.replay.review.domain.ReviewImage;
