+++ src/main/java/phanes/replay/review/service/ReviewService.java
@@ -9,13 +9,13 @@ import phanes.replay.exception.ReviewNotFountException;
 import phanes.replay.image.service.S3Service;
 import phanes.replay.review.domain.Review;
 import phanes.replay.review.domain.ReviewImage;
+import phanes.replay.review.dto.mapper.ReviewMapper;
 import phanes.replay.review.dto.request.ReviewCreateRq;
 import phanes.replay.review.dto.request.ReviewUpdateRq;
 import phanes.replay.review.dto.response.ReviewRatingRs;
 import phanes.replay.review.dto.response.ReviewRs;
-import phanes.replay.review.mapper.ReviewMapper;
-import phanes.replay.review.repository.ReviewImageRepository;
-import phanes.replay.review.repository.ReviewRepository;
+import phanes.replay.review.persistence.repository.ReviewImageRepository;
+import phanes.replay.review.persistence.repository.ReviewRepository;
 import phanes.replay.theme.domain.Theme;
 import phanes.replay.theme.service.ThemeQueryService;
 import phanes.replay.user.domain.User;
